Add code to post data to "How out of date are we?"
Unlike the existing [HOODAW updater], this expects
an env var containing the HOODAW API key.

Although this adds a new dependency, it means we
don't need to add kubectl to this docker image, or
cluster credentials. 

In general, I think moving to separate jobs to
post data to HOODAW is a more sensible approach
than the current one of adding more and more
report types to a single updater script in the 
HOODAW updater docker image.


[HOODAW updater]: https://github.com/ministryofjustice/cloud-platform-how-out-of-date-are-we/blob/main/updater-image/update.sh