Exclude security groups from orphaned AWS list
Security groups are created by some of the
components of the cloud platform, and are often
*not mentioned* in any `terraform.tfstate` files.

So, listing them in the "orphaned AWS resources"
page makes it likely that we will delete something
important, by accident.

The simplest solution is to not list any security
groups at all, until we have a more reliable way
to identify security groups which are definitely
not useful (check to see if the VPC they relate
to still exists? Do security groups persist if
the VPC they relate to is deleted?)